// src/app/projects/models/project-area.model.ts
export class ProjectArea {
    projectAreaId: string;
    areaId: string;
    projectEquipmentId: string;
    name: string;
    elevationUnit: number;
    elevationStart: number;
    elevationEnd: number;
    elevationIncrement: number;
    tubeDiameter: number;
    readings: number;
    tubesStart: number;
    tubesEnd: number;
    referenceTube: number;
    referenceElevation: number;
    referenceX: string;
    referenceY: string;
    filename: string;
    createdAt: string;
    updatedAt: string;
}